agregar guards

+++ src/app/category/category.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { Subscription } from 'rxjs';
+import { TokenService } from '../auth/service/token.service';
 import { ModalCategoryComponent } from '../modals/modal-category/modal-category.component';
 import { CategoryFecthBook } from '../models/category';
 import { CategoryService } from '../service/category-service.service';
@@ -10,7 +11,7 @@ import { CategoryService } from '../service/category-service.service';
   templateUrl: './category.component.html',
   styleUrls: ['./category.component.css']
 })
-export class CategoryComponent implements OnInit, OnDestroy{
+export class CategoryComponent implements OnInit, OnDestroy {
 
 
   private subscription: Subscription = new Subscription();
@@ -18,18 +19,21 @@ export class CategoryComponent implements OnInit, OnDestroy{
 
   listCategoryResp: CategoryFecthBook[] = [];
 
+  isAdmin = false;
 
   constructor(private categoryService: CategoryService,
+    private tokenService: TokenService,
     private modalService: NgbModal) { }
-    
+
 
   ngOnInit(): void {
 
+    this.isAdmin = this.tokenService.isAdmin();
     this.findAllCategory();
 
-    this.subscription.add(  
+    this.subscription.add(
 
-      this.categoryService.refresh.subscribe(resp =>{
+      this.categoryService.refresh.subscribe(resp => {
 
         this.findAllCategory();
       })
@@ -39,13 +43,13 @@ export class CategoryComponent implements OnInit, OnDestroy{
 
 
   ngOnDestroy(): void {
-   
+
     this.subscription.unsubscribe();
   }
 
-  private findAllCategory(){
+  private findAllCategory() {
 
-    this.categoryService.findlFetchBookAllCategory().subscribe(resp =>{
+    this.categoryService.findlFetchBookAllCategory().subscribe(resp => {
 
       this.listCategoryResp = resp.data;
 
@@ -54,26 +58,31 @@ export class CategoryComponent implements OnInit, OnDestroy{
 
   }
   deleteCategory(ide: number) {
-  
-    this.categoryService.deleteCategory(ide).subscribe(resp =>{
 
-      console.log("categoria eliminada");
-    },error =>{
-      console.log("Error categoria eliminar", error);
 
-    });
+
+    if (confirm("Seguro desea eliminar la categoria?")) {
+      this.categoryService.deleteCategory(ide).subscribe(resp => {
+
+        console.log("categoria eliminada");
+      }, error => {
+        console.log("Error categoria eliminar", error);
+
+      });
+    }
+
 
 
   }
   openModalUpdate(ide: number) {
-    const currentModal= this.modalService.open(ModalCategoryComponent, {size: 'lg'});
+    const currentModal = this.modalService.open(ModalCategoryComponent, { size: 'lg' });
     currentModal.componentInstance.ideCategoryUpdate = ide;
   }
 
 
   openModal() {
 
-    this.modalService.open(ModalCategoryComponent, {size: 'lg'});
+    this.modalService.open(ModalCategoryComponent, { size: 'lg' });
 
   }
