+++ src/app/menu/menu.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { Subscription } from 'rxjs';
+import { TokenService } from '../auth/service/token.service';
 import { ModalBookComponent } from '../modals/modal-book/modal-book.component';
 import { BookResponseDto } from '../models/dto-response/book-response-dto';
 import { BookService } from '../service/book-service.service';
@@ -12,16 +13,20 @@ import { BookService } from '../service/book-service.service';
 })
 export class MenuComponent implements OnInit, OnDestroy{
 
+  isAdmin = false;
   public listBookResponse : BookResponseDto[] = [];
 
   private subscription: Subscription = new Subscription();
 
 
   constructor(private bookService: BookService,
-    private modalService: NgbModal) { }
+    private modalService: NgbModal,
+    private tokenService: TokenService) { }
   
 
   ngOnInit(): void {
+
+    this.isAdmin = this.tokenService.isAdmin();
 
     this.findAllBook();
 
@@ -54,13 +59,18 @@ export class MenuComponent implements OnInit, OnDestroy{
   deleteBook(ide:number){
 
 
-    this.bookService.deleteBook(ide).subscribe(resp =>{
-
-       console.log("Book deleted");
-    },
-    error =>{
-      console.log(error);
-    });
+    let response = confirm("Seguro desea eliminar el libro");
+    
+    if (response){
+      
+      this.bookService.deleteBook(ide).subscribe(resp =>{
+        
+           console.log("Book deleted");
+        },
+        error =>{
+            console.log(error);
+          });
+        }
   
 }
 openModalUpdate(ide:number){
